+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 Upgrading to 3.8.x versions requires that you upgrade to latest 3.5.x version first.
 
+- Fix Custom Field Options order not being respected, #702, #703
+
 [3.8.4]: https://github.com/eventum/eventum/compare/v3.8.3...master
 
 ## [3.8.3] - 2019-09-29